EOS Transport "clientId" => "transportId", associated comments
This commit and PR is focused on addressing a confusion between NetworkManager's "client id" and NetworkTransport's "transport id" concept. This is meant to aid in determining how to manage a client's network connection by providing a way to identify their transport id.

When a user connects to a host, they are assigned a transport id by the NetworkTransport. Then they are assigned a client id by NetworkManager. These can be similar numbers, but aren't actually related. Instead, whenever NetworkManager needs to interface to and from NetworkTransport, it looks up the transport id/client id as appropriate.

The transport id is stable between disconnections, and EOSTransport will recognize previously connected clients. You can always depend on the EOSTransport to be able to provide you a transport based on the ProductUserId.

This PR aims to add a lot of clarification around these concepts and names. There's no functionality change in this pr, just comments and name changes.