feat: Notify user that Ecommerce won't work in the Editor
Some features of the plugin, such as the Ecommerce interface, require the use of the Social Overlay. If the Social Overlay isn't available, it won't work. This aims to cover exactly one scenario: the Unity Editor. If UNITY_EDITOR is defined, `DisabledInEditorWarningPanel` will enable a game object. That game object may contain text that explains the error. There is no standardized warning panel prefab, and should be used as-needed.

#EOS-2130