fix: Parse Enums as Integers for Comparison in `ListOfStringsToEnumConverters.FromNumberValue`
`ListOfStringsToEnumConverters.FromNumberValue` converts a string to an enum, and then compares that enum to find the "lowest" value. If the enum fails to parse, it will set itself properly to the lowest value. This was trying to compare an Enum to an integer, and was previously failing when the integer was 0 but the Enum didn't have a "0" defined. This fix explicitly compares the underlying value instead of comparing the enum, which covers this scenario.